move util function and hook to separate files
